requirements.txt
aiogram
magic-filter
asyncpg
python-dotenv